+++ RecatchLegendaryFish/release-notes.md
@@ -1,5 +1,7 @@
 ﻿# Release notes
-## Upcoming release
+## 1.7
+Released 26 March 2021 for SMAPI 3.9.5 or later. Updated by Pathoschild.
+
 * Added optional key to toggle recatch mode.
 * Updated for split-screen mode in Stardew Valley 1.5.
 * Fixed compatibility with [unofficial 64-bit mode](https://stardewvalleywiki.com/Modding:Migrate_to_64-bit_on_Windows).